+++ lib/PlatformBaseModel.php
@@ -42,6 +42,14 @@ abstract class PlatformBaseModel
     /**
      * @return static
      */
+    public static function createSlave()
+    {
+        return self::create(GnfConnectionProvider::getConnection(PlatformConnectionGroup::PLATFORM_SLAVE));
+    }
+
+    /**
+     * @return static
+     */
     public static function createPlatformOnlyWrite()
     {
         return self::create(GnfConnectionProvider::getConnection(PlatformConnectionGroup::PLATFORM_ONLY_DB_WRITE));
@@ -53,6 +61,14 @@ abstract class PlatformBaseModel
     public static function createPlatformOnlyRead()
     {
         return self::create(GnfConnectionProvider::getConnection(PlatformConnectionGroup::PLATFORM_ONLY_DB_READ));
+    }
+
+    /**
+     * @return static
+     */
+    public static function createPlatformOnlySlave()
+    {
+        return self::create(GnfConnectionProvider::getConnection(PlatformConnectionGroup::PLATFORM_ONLY_DB_SLAVE));
     }
 
     /**
@@ -74,6 +90,14 @@ abstract class PlatformBaseModel
     /**
      * @return static
      */
+    public static function createPlatformBookSlave()
+    {
+        return self::create(GnfConnectionProvider::getConnection(PlatformConnectionGroup::PLATFORM_BOOK_DB_SLAVE));
+    }
+
+    /**
+     * @return static
+     */
     public static function createCpstatWrite()
     {
         return self::create(GnfConnectionProvider::getConnection(PlatformConnectionGroup::CP_STATISTICS));
@@ -85,6 +109,14 @@ abstract class PlatformBaseModel
     public static function createCpstatRead()
     {
         return self::create(GnfConnectionProvider::getConnection(PlatformConnectionGroup::CP_STATISTICS_READ));
+    }
+
+    /**
+     * @return static
+     */
+    public static function createCpstatSlave()
+    {
+        return self::create(GnfConnectionProvider::getConnection(PlatformConnectionGroup::CP_STATISTICS_SLAVE));
     }
 
     public function transactional(callable $callable)