+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.3] - 2018-10-16
+### ADDED
+- add PLATFORM_SLAVE to PlatformConnectionGroup
+- add PLATFORM_ONLY_DB_SLAVE to PlatformConnectionGroup
+- add PLATFORM_BOOK_DB_SLAVE to PlatformConnectionGroup
+- add CP_STATISTICS_SLAVE to PlatformConnectionGroup
+
 ## [0.2.2] - 2018-10-16
 ### CHANGED
 - Change composer dependency: platform-gnfdb(^0.1.11 -> ^0.1.12) for using sqlInsertOrUpdateBulk